+++ src/main/java/pl/plajer/villagedefense/arena/ArenaManager.java
@@ -175,7 +175,7 @@ public class ArenaManager {
       return false;
     }
 
-    if (!plugin.getConfigPreferences().getOption(ConfigPreferences.Option.BUNGEE_ENABLED)) {
+    if (!plugin.getModuleLoader().isModulePresent("Bungee Cord")) {
       String perm = PermissionsManager.getJoinPerm();
       if (!(player.hasPermission(perm.replace("<arena>", "*")) || player.hasPermission(perm.replace("<arena>", arena.getId())))) {
         player.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage(Messages.JOIN_NO_PERMISSION)