+++ src/main/java/pl/plajer/villagedefense/commands/arguments/admin/ModulesArgument.java
@@ -1,0 +1,48 @@
+/*
+ * Village Defense - Protect villagers from hordes of zombies
+ * Copyright (C) 2019  Plajer's Lair - maintained by Plajer and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package pl.plajer.villagedefense.commands.arguments.admin;
+
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
+import pl.plajer.villagedefense.commands.arguments.ArgumentsRegistry;
+import pl.plajer.villagedefense.commands.arguments.data.CommandArgument;
+import pl.plajer.villagedefense.commands.arguments.data.LabelData;
+import pl.plajer.villagedefense.commands.arguments.data.LabeledCommandArgument;
+import pl.plajer.villagedefense.handlers.module.ModuleVisualizer;
+
+/**
+ * @author Plajer
+ * <p>
+ * Created at 17.06.2019
+ */
+public class ModulesArgument {
+
+  public ModulesArgument(ArgumentsRegistry registry) {
+    registry.mapArgument("villagedefenseadmin", new LabeledCommandArgument("modules", "villagedefense.admin.modules", CommandArgument.ExecutorType.PLAYER,
+        new LabelData("/vda modules", "/vda modules",
+            "&7Displays GUI with all Village Defense modules\n&6Permission: &7villagedefense.admin.modules")) {
+      @Override
+      public void execute(CommandSender sender, String[] args) {
+        new ModuleVisualizer().openInventory((Player) sender);
+      }
+    });
+  }
+
+}