+++ CHANGELOG.md
@@ -9,11 +9,16 @@ deploy webhooks to print changelog successfully
 
 That's all, matcher will stop when detects next line started with `###` match
 
-### 4.2.2 Release (14/15.06.2019)
+### 4.2.2 Release (14/18.06.2019)
 * Fixed setup menu completed/not completed values were buggy
 * Fixed /vda setprice message was sent twice
 * Now shop will count wolf/golem spawn items to total price spent in game in game summary
 * Join permission message outputs required permission node now
+* Fixed bungee failed to add players after the game if shutdown on game finish was disabled
+* Implemented modules - more info needed
+* Stripped plugin from BungeeCord in favor of Bungee Cord Module
+* Stripped plugin from /vd randomjoin command in favor of RandomJoin Module
+* Stripped plugin from Wolves/Golems upgrades in favor of Entities Upgrades Module
 
 ### 4.2.1 Release (25.05.2019 - 08.06.2019)
 * Fixed possible issues with scoreboard colors